+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 # # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## 2.0.0 2016-09-06
+
+### Changed
+- json-schema-object callback to override required + properties
 
 ## 1.1.0 2016-09-04
 
 ### Added
-- pattern to JSONString 
+- pattern to JSONString
 - additionalProperties to JSONObject
-
 
 ## 1.0.0 2016-09-03